add printff fprmat rulr and adjust print_nbr bug

+++ libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: eszu <eszu@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/03 15:55:31 by pruenrua          #+#    #+#             */
-/*   Updated: 2022/09/07 01:18:18 by eszu             ###   ########.fr       */
+/*   Updated: 2022/09/07 22:06:53 by eszu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,12 +15,13 @@
 # include <stdlib.h>
 # include <unistd.h>
 # include <stdarg.h>
+# include <stdio.h>
 
 int		ft_printf(const char *, ...);
 int     ft_strlen_f(char *str);
 int     ft_putchar_f(int c);
 int     ft_putstring_f(char *str);
-int     ft_print_nbr_base(long value, char *base);
+int     ft_print_nbr_base(int value, char *base);
 int     ft_print_nbr_base_u(unsigned long value, char *base);
 
 #endif
